backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
email-validator
python-jose[cryptography]
passlib[bcrypt]
requests
python-multipart
